backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1

# Groq + compatible httpx
groq==0.9.0
httpx==0.27.0